Decoding: Decoding works

+++ app/src/main/java/nl/hva/vuwearable/models/Measurement.kt
@@ -3,7 +3,7 @@ package nl.hva.vuwearable.models
 data class Measurement(
     val totalBytes: Int,
     val title: String,
-    val formula: ((Long) -> Double)? = null
+    val formula: ((Int) -> Double)? = null
 ) {
-    var value = 0L
+    var value = 0.0
 }
